chore: removed MediaLibraryContext
