+++ app/app.component.tsx
@@ -10,22 +10,20 @@ import {Routes} from './constants';
 
 function AppComponent() {
   return (
-    <AppContexts.MediaLibraryProvider>
-      <AppContexts.MediaPlayerProvider>
-        <MediaSidebarComponent/>
-        <MediaPlayerRibbonComponent/>
-        <BrowserRouter>
-          <Switch>
-            <Route
-              path={Routes.HOME}
-              render={() => (
-                <AppPages.HomeComponent/>
-              )}
-            />
-          </Switch>
-        </BrowserRouter>
-      </AppContexts.MediaPlayerProvider>
-    </AppContexts.MediaLibraryProvider>
+    <AppContexts.MediaPlayerProvider>
+      <MediaSidebarComponent/>
+      <MediaPlayerRibbonComponent/>
+      <BrowserRouter>
+        <Switch>
+          <Route
+            path={Routes.HOME}
+            render={() => (
+              <AppPages.HomeComponent/>
+            )}
+          />
+        </Switch>
+      </BrowserRouter>
+    </AppContexts.MediaPlayerProvider>
   );
 }
