+++ app/contexts/index.ts
@@ -1,9 +1,4 @@
 export {
-  MediaLibraryContext,
-  MediaLibraryProvider,
-} from './media-library.context';
-
-export {
   MediaPlayerContext,
   MediaPlayerProvider,
 } from './media-player.context';
