+++ app/services/index.ts
@@ -1,9 +1,11 @@
 import I18nService from './i18n.service';
 import MediaService from './media.service';
+import MediaLibraryService from './media-library.service';
 import SystemService from './system.service';
 
 export {
   I18nService,
   MediaService,
+  MediaLibraryService,
   SystemService,
 };
